refactor account details section to AccountDetailsBox component
