+++ client/src/pages/Account/index.js
@@ -1,1 +1,2 @@
 export { default as AccountSettings } from './AccountSettings';
+export { default as AccountDetailsBox } from './AccountDetailsBox';
